Hubspot: Update modules on entry save [INTEG-2785] (#9975)
* first version of update on save event + some tests

* adding refactor to the appEventHandler tests

* rename from test

* Fix linter

* changin place of call and retry

* refactor: passing only the hubspot token to function

* early return if entry is not synced + new test

* fix test

---------

+++ apps/hubspot/functions/appEventHandler.ts
@@ -5,7 +5,9 @@ import {
   AppEventRequest,
 } from '@contentful/node-apps-toolkit';
 import ConfigEntryService from '../src/utils/ConfigEntryService';
-import { initContentfulManagementClient } from './common';
+import { createModuleFile, getFiles, initContentfulManagementClient } from './common';
+import { AppInstallationParameters, ConnectedField } from '../src/utils/utils';
+import { PlainClientAPI } from 'contentful-management';
 
 const WAIT_TIMES = [1000, 2000];
 
@@ -22,8 +24,78 @@ export const handler: FunctionEventHandler<FunctionTypeEnum.AppEventHandler> = a
 
   const configService = new ConfigEntryService(cma);
 
+  const entryConnectedFields = await configService.getEntryConnectedFields(entryId);
+  if (!entryConnectedFields || entryConnectedFields.length === 0) {
+    // Entry is not tracked in config, skip the function
+    return;
+  }
+
   if (contentfulTopic.includes('Entry.delete')) {
     await callAndRetry(() => configService.removeEntryConnectedFields(entryId));
+  } else if (
+    contentfulTopic.includes('Entry.save') ||
+    contentfulTopic.includes('Entry.auto_save')
+  ) {
+    const { hubspotAccessToken } = context.appInstallationParameters as AppInstallationParameters;
+    await callAndRetry(() => entrySavedHandler(cma, hubspotAccessToken, body, configService));
+  }
+};
+
+const entrySavedHandler = async (
+  cma: PlainClientAPI,
+  hubspotAccessToken: string,
+  body: any,
+  configService: ConfigEntryService
+) => {
+  const entryConnectedFields = await configService.getEntryConnectedFields(body.sys.id);
+
+  const contentType = await cma.contentType.get({
+    contentTypeId: body.sys.contentType.sys.id,
+  });
+
+  const results: Array<ConnectedField> = [];
+  for (const connectedField of entryConnectedFields) {
+    const field = body.fields[connectedField.fieldId];
+    const locale = connectedField.locale || Object.keys(field)[0];
+    let fieldValue = field[locale];
+    const fieldInfo = contentType.fields.find(
+      (f: { id: string }) => f.id === connectedField.fieldId
+    );
+
+    if (!fieldValue || !fieldInfo) {
+      continue;
+    }
+
+    let updateResult: ConnectedField = {
+      fieldId: connectedField.fieldId,
+      locale,
+      moduleName: connectedField.moduleName,
+      updatedAt: new Date().toISOString(),
+    };
+
+    try {
+      const { fieldsFile } = getFiles(fieldInfo.type, fieldValue);
+      await createModuleFile(
+        fieldsFile,
+        'fields.json',
+        connectedField.moduleName,
+        hubspotAccessToken
+      );
+    } catch (error: any) {
+      updateResult = {
+        ...updateResult,
+        error: {
+          status: error?.statusCode || 500,
+          message: error?.message || String(error),
+        },
+      };
+    }
+
+    results.push(updateResult);
+  }
+
+  if (results.length > 0) {
+    await configService.updateEntryConnectedFields(body.sys.id, results);
   }
 };
